% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecalFunctions.R
\name{AddCalculatedSummary}
\alias{AddCalculatedSummary}
\title{AddCalculatedSummary function}
\usage{
AddCalculatedSummary(Recal)
}
\arguments{
\item{Recal}{data frame}
}
\value{
A modified data frame with added calculated summary columns.
}
\description{
This function takes a data frame 'Recal' and calculates summary statistics
grouped by the 'Index' column. It adds columns for Minimum, Maximum, Mean,
Maximum Mass, Maximum Intensity, Second Maximum Intensity, and Second Maximum Mass.
}
