% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{bind_and_filter_data}
\alias{bind_and_filter_data}
\title{Binds and filters two data frames}
\usage{
bind_and_filter_data(data1, data2, remove_na)
}
\arguments{
\item{data1}{data frame:
The first data frame to be combined.}

\item{data2}{data frame:
The second data frame to be combined.}

\item{remove_na}{logical:
If TRUE, remove rows with missing experimental masses or masses less than zero.}
}
\value{
data frame:
The combined and filtered data frame.
}
\description{
This function combines two data frames and removes duplicate rows. Optionally, it can remove rows with missing
experimental masses or masses less than or equal to zero.
}
