% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{calculate_variables}
\alias{calculate_variables}
\title{Calculate variables based on mass values}
\usage{
calculate_variables(
  exp_mass_col,
  mass_difference,
  normalization_factor,
  multiplier
)
}
\arguments{
\item{exp_mass_col}{numeric:
experimental mass values.}

\item{mass_difference}{numeric:
Mass difference value.}

\item{normalization_factor}{numeric:
Normalization factor.}

\item{multiplier}{numeric:
Multiplier value.}
}
\value{
matrix:
calculated variables: km, kmd, kmr, and kmdr.
}
\description{
This function calculates variables based on specified mass values and parameters.
}
