% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{combine_isotopes}
\alias{combine_isotopes}
\title{Combines isotopes
This function combines isotopes based on experimental mass.}
\usage{
combine_isotopes(Isotopes)
}
\arguments{
\item{Isotopes}{data frame:
The input data frame containing isotopes.}
}
\value{
data frame:
The combined data frame.
}
\description{
Combines isotopes
This function combines isotopes based on experimental mass.
}
