% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{compare_bad_and_filter}
\alias{compare_bad_and_filter}
\title{Compare and filter isotopic and bad data}
\usage{
compare_bad_and_filter(iso_df, bad_df)
}
\arguments{
\item{iso_df}{data frame:
Data frame containing isotopic data.}

\item{bad_df}{data frame:
Data frame containing bad data for comparison.}
}
\value{
list:
A list containing filtered isotopic and bad data frames.
}
\description{
This function compares isotopic and bad data frames based on mass, abundance, and retention time.
It returns a list containing filtered isotopic and bad data frames.
}
