% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{CompIntToFactor}
\alias{CompIntToFactor}
\title{Component integer to factor}
\usage{
CompIntToFactor(x)
}
\arguments{
\item{x}{component; numeric; integer representation for component}
}
\value{
factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Fl",
 "Cl37", "M", "NH4", "POE", "NOE", "Z"})
}
\description{
Given a component as an integer returns the correponding factor representation.
For example, 1 -> "C", 2 -> "H", ..., 9 -> "Z"
}
\details{
This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
CompIntToFactor(1)

CompIntToFactor(2)

}
