% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{compute_combinations}
\alias{compute_combinations}
\title{Computes all possible combinations of series
This function computes all possible combinations of series. The number of combinations is specified by user.}
\usage{
compute_combinations(df, n)
}
\arguments{
\item{df}{DataFrame A pre-filtered dataframe containing series.}

\item{n}{Integer Number of combinations to compute.}
}
\value{
DataFrame A dataframe of row-index based combinations.
}
\description{
Computes all possible combinations of series
This function computes all possible combinations of series. The number of combinations is specified by user.
}
