% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{compute_coverage}
\alias{compute_coverage}
\title{Computes coverage of a subset.
This function computes m/z range coverage of a particular subset. The higher coverage, the better.}
\usage{
compute_coverage(subset, global_max, global_min)
}
\arguments{
\item{subset}{DataFrame A subset of pre-filtered input data, has a size of combination number.}

\item{global_max}{Float A higher bound of the instrument m/z range.}

\item{global_min}{Float A lower bound of the instrument m/z range.}
}
\value{
Coverage (in \%) of a particular subset.
}
\description{
Computes coverage of a subset.
This function computes m/z range coverage of a particular subset. The higher coverage, the better.
}
