% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{compute_final_score}
\alias{compute_final_score}
\title{Compute a sum score for the whole combination.
This function computes a score summing all parameter scores and arranges them in descending manner.}
\usage{
compute_final_score(scores_df)
}
\arguments{
\item{scores_df}{DataFrame A dataframe of scores.}
}
\value{
DataFrame A sorted dataframe containing the final score.
}
\description{
Compute a sum score for the whole combination.
This function computes a score summing all parameter scores and arranges them in descending manner.
}
