% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{compute_pairs}
\alias{compute_pairs}
\title{Compute pairs based on specified criteria and filtering}
\usage{
compute_pairs(
  data,
  err,
  mass_difference,
  normalization_factor,
  multiplier,
  pair_thresold,
  filter_criteria
)
}
\arguments{
\item{data}{data frame:
Data frame containing experimental mass values.}

\item{err}{numeric:
Error value for filtering.}

\item{mass_difference}{numeric:
Mass difference value.}

\item{normalization_factor}{numeric:
Normalization factor.}

\item{multiplier}{numeric:
Multiplier value.}

\item{pair_thresold}{numeric vector:
Vector containing pair mass difference criteria.}

\item{filter_criteria}{numeric vector:
Vector containing filtering criteria.}
}
\value{
data frame:
A data frame containing computed pairs of mass values that satisfy the specified criteria and filtering parameters.
}
\description{
This function computes pairs of mass values based on specified criteria and filtering parameters.
}
