% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{compute_scores}
\alias{compute_scores}
\title{Computes the scores for individual subsets.
This function computes individual scores for the particular combination of series.}
\usage{
compute_scores(combination)
}
\arguments{
\item{combination}{DataFrame A subset of pre-filtered input data, has a size of combination number.}
}
\value{
List List of scores for individual parameters.
}
\description{
Computes the scores for individual subsets.
This function computes individual scores for the particular combination of series.
}
