% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{compute_subsets}
\alias{compute_subsets}
\title{Computes subsets based on combinations.
This function selects subsets of the input data based on the combinations of series.}
\usage{
compute_subsets(df, n)
}
\arguments{
\item{df}{DataFrame A pre-filtered dataframe containing series.}

\item{n}{Integer Number of combinations to compute.}
}
\value{
A list of subsets based on series combinations.
}
\description{
Computes subsets based on combinations.
This function selects subsets of the input data based on the combinations of series.
}
