% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{create_data_chunks}
\alias{create_data_chunks}
\title{Creates data chunks}
\usage{
create_data_chunks(data1)
}
\arguments{
\item{data1}{data frame:
The input data frame to be divided into chunks.}
}
\value{
list:
A list containing two elements: "Carblist" and "Sulflist," each representing a list of data chunks.
}
\description{
This function takes a data frame and divides it into chunks.
}
