% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{create_pairs}
\alias{create_pairs}
\title{Create pairs based on mass difference criteria}
\usage{
create_pairs(exp_mass, less_than, greater_than)
}
\arguments{
\item{exp_mass}{numeric vector:
Vector of experimental mass values.}

\item{less_than}{numeric:
pair thresold.}

\item{greater_than}{numeric:
pair thresold.}
}
\value{
data frame:
A data frame containing pairs of experimental mass values that satisfy the specified mass difference criteria.
}
\description{
This function generates pairs of mass values based on specified criteria.
}
