% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Halogen.R
\name{EM2}
\alias{EM2}
\title{Expected masses of components}
\usage{
EM2(x)
}
\arguments{
\item{x}{component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})}
}
\value{
numeric; value of expected mass
}
\description{
Returns the nominal masses of molecular components, using the mass of an
  electron for charge
}
\details{
This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
EM("C")

EM('C')

}
