% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R, R/Utilities_Halogen.R
\name{Even}
\alias{Even}
\title{Even integer}
\usage{
Even(x)

Even(x)
}
\arguments{
\item{x}{numeric}
}
\value{
Boolean

Boolean
}
\description{
Returns Boolean if number is Even

Returns Boolean if number is Even
}
\examples{
Even(123)

Even(12)

Even(123)

Even(12)

}
