% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{extract_mono_and_iso_pairs}
\alias{extract_mono_and_iso_pairs}
\title{Extracts mono and iso pairs from a data frame}
\usage{
extract_mono_and_iso_pairs(pairs)
}
\arguments{
\item{pairs}{data frame:
The input data frame}
}
\value{
list:
A list with two elements: "mono_pair" containing columns "Mono_mass" and "Iso_mass1,"
and "iso_pair" containing columns "Iso_mass1" and "Iso_mass2."
}
\description{
This function takes a data frame of pairs and extracts columns for mono and iso masses.
}
