% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{filter_pairs}
\alias{filter_pairs}
\title{Filter pairs based on mass difference error}
\usage{
filter_pairs(pairs_data, mass_difference, threshold)
}
\arguments{
\item{pairs_data}{data frame:
Data frame containing pairs of mass values.}

\item{mass_difference}{numeric:
Mass difference value.}

\item{threshold}{numeric:
Error threshold for filtering pairs.}
}
\value{
data frame:
A data frame containing pairs of mass values that satisfy the specified error threshold.
}
\description{
This function filters pairs based on mass difference error.
}
