% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{filter_recal_series}
\alias{filter_recal_series}
\title{Filters the input dataframe
This function filters the input dataframe based on abundance score threshold and peak distance threshold; and
computes the length of the series.}
\usage{
filter_recal_series(df, abundance_score_threshold, peak_distance_threshold)
}
\arguments{
\item{df}{DataFrame An output from RecalList, containing recalibrant CH2 series.}

\item{abundance_score_threshold}{Float A threshold for filtering abundance score parameter. The series with higher values #' are better. Default value is 100.}

\item{peak_distance_threshold}{Float A threshold for the peak distance parameter. The closer this value is to 1, the
better.}
}
\value{
DataFrame A filtered dataframe.
}
\description{
Filters the input dataframe
This function filters the input dataframe based on abundance score threshold and peak distance threshold; and
computes the length of the series.
}
