% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{filter_subsets_based_on_coverage}
\alias{filter_subsets_based_on_coverage}
\title{Filters subsets based on the coverage.
This function filters the subsets based on the coverage threshold.}
\usage{
filter_subsets_based_on_coverage(
  subsets,
  coverage_threshold,
  global_max,
  global_min
)
}
\arguments{
\item{subsets}{DataFrame A subset of pre-filtered input data, has a size of combination number.}

\item{coverage_threshold}{Integer How many \% of the m/z range should be covered. Default is 90 \%.}

\item{global_max}{Float A higher bound of the instrument m/z range.}

\item{global_min}{Float A lower bound of the instrument m/z range.}
}
\value{
List A list of subsets which pass the coverage threshold (their coverage is higher than the threshold)
}
\description{
Filters subsets based on the coverage.
This function filters the subsets based on the coverage threshold.
}
