% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{filtered_data_is_empty}
\alias{filtered_data_is_empty}
\title{Return TRUE if any of the elements in the list has a length of 0, otherwise FALSE}
\usage{
filtered_data_is_empty(filtered_data)
}
\arguments{
\item{filtered_data}{Filtered data with mono-iso links.}
}
\value{
TRUE if any of the links is 0, otherwise FALSE.
}
\description{
Return TRUE if any of the elements in the list has a length of 0, otherwise FALSE
}
