% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRecalSeries.R
\name{find_final_series}
\alias{find_final_series}
\title{Selects final series.
This function selects final best-scoring series. An user can choose, whether only a number of series corresponding
to the size of the combination will be returned, or 10 best-scoring 10 series will be returned.}
\usage{
find_final_series(scores_df, number_of_combinations, fill_series)
}
\arguments{
\item{scores_df}{DataFrame A dataframe of scores.}

\item{number_of_combinations}{Integer Number of combinations to compute.}

\item{fill_series}{Boolean If TRUE, top 10 unique series will be returned, otherwise only the series from the best
combination will be returned.}
}
\value{
DataFrame A dataframe of best-scoring series.
}
\description{
Selects final series.
This function selects final best-scoring series. An user can choose, whether only a number of series corresponding
to the size of the combination will be returned, or 10 best-scoring 10 series will be returned.
}
