% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCoreFormulae2_Halogen.R
\name{FindCoreFormulae2_Halo}
\alias{FindCoreFormulae2_Halo}
\title{Assigns all possible combinations of C, H, and O for a given MW}
\usage{
FindCoreFormulae2_Halo(env)
}
\arguments{
\item{env}{environment from \code{\link{MFAssign}} function}
}
\description{
FindCoreFormulae assigns all possible combinations of C, H, and O for a
given MW and subject to constraints in the \code{\link{ValidFormula}} function.
This is a subfunction of the MFAssign function and will not work
independently of that function.
}
\examples{
env <- environment()
FindCoreFormulae(env)

}
