% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{HighMoles}
\alias{HighMoles}
\title{Maximum limit of moles (user limit)}
\usage{
HighMoles(x, N = 0, S = 0, P = 0, Cl = 0, Fl = 0, POE = 0,
  NOE = 0, E = 0, S34 = 0, Cl37 = 0, N15 = 0, D = 0, M = 0,
  NH4 = 0, Z = 1)
}
\arguments{
\item{x}{component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})}
}
\value{
numeric; value of maximum limit of moles
}
\description{
Returns the maximum limit of moles used by the
  \code{\link{MFAssign}} function, specified by the user.
}
\details{
This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
HighMoles("C")

HighMoles('C')

}
