% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{IsoFiltR}
\alias{IsoFiltR}
\title{Identifies and separates likely isotopic masses from monoisotopic masses}
\usage{
IsoFiltR(peaks, SN = 0, Diffrat = 0.1, Sulfrat = 30)
}
\arguments{
\item{peaks}{data frame:
The input 2 column data frame containing abundance and peak mass}
}
\value{
list(Monolist, Isolist):
  Monolist - monoistopic and non-matched masses,
  Isolist - isotopic masses
}
\description{
IsoFiltR separates likely isotopic masses from monoisotopic masses in a
mass list. It can identify likely 13C isotopic
masses and put them in a separate mass list from the monoisotopic
masses. This should be done prior to formula assignment in order
to lessen the chances of incorrectly assigned formulas, where, for
example a 13C containing CHO formula can be assigned to a
monoistopic CHNO formula.
}
\details{
The only necessary input is a two column data frame with the abundance in the
first column and the measured ion mass in the second column. This should be the
raw mass list output.

At this time only the 13C isotope can be identified, although plans to improve
the number of isotopes to be available are in process. The output of this function
is a list of dataframes. Dataframe 1 contains the flagged monoisotopic masses, and
all masses that did not have a matching isotopic mass. The second dataframe contains
the masses flagged as isotopic. These two dataframes should next be run in the
 \code{\link{MFAssignCHO}} or \code{\link{MFAssign}} to assign formulas to the masses.

Note that the classification of isotopic or monoisotopic from this function is not
definitive.
}
\examples{
IsoFiltR(peaks=df)
}
