% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMDNoise.R
\name{KMDNoise}
\alias{KMDNoise}
\title{Estimates the Signal to Noise cut}
\usage{
KMDNoise(df, upper = 0.2, lower = 0.05)
}
\arguments{
\item{df}{- dataframe of intensity and ion mass, column 1 should be mass, column 2 should be intensity}

\item{upper}{- sets the upper limit for the y intercept, default is 0.2}

\item{lower}{- sets the lower limit for the y intercept, default is 0.05}
}
\value{
list(Noise = Noise, Plot = KMD)
        Noise - numeric value of the noise level for the data set
        Plot - plot showing the KMD for the inputted data with the
        noise estimation area highlighted.
}
\description{
For raw mass spectral data using a method
based on the separation of Kendrick mass
defect of intense analyte peaks and low
intensity noise peaks.
}
\details{
When the KMD is calculated for all peaks
in a raw mass spectrum, the more intense
analyte peaks form "islands" surrounded
by a sea of low intensity noise peaks.
using the mathematical limits of the KMD
calculation the slope of a KMD plot can be
calculated and used in conjunction with the
KMD limits of chemically feasible molecular
formulas to isolate regions of the plot that
contain only noise peaks. These noise peaks can
then be averaged to determine the signal to noise
of the spectrum.

The upper chemical limit of the KMD plot can be
estimated by calculating the KMD for non-hydrogenated
condensed carbon formulas (C10, C20, C30, etc.).
This provides an x,y intercept of (0,0) when the slope
(0.001232) of the KMD plot is used. This provides a
reasonable lower bound for isolating noise peaks.
The upper bound can be selected by increasing the
intercept value in the equation for the KMD plot,
which is y = 0.0011232 * x + b.

This signal to noise estimation method uses the
equation described above to select a region of the
KMD plot that contains only noise and then estimates
the average intensity of the peaks in that region,
reporting this value as the noise level.

An additional output of the function is the KMD plot
with the bounds of the noise estimation area highlighted
in red. This noise level can then be multiplied by the
users chosen value (3, 6, 10) in order to set the
signal to noise cut for formula assignment.

The y intercept for the upper and lower bounds are set
to 0.05 (lower) and 0.2 (upper). The lower bound is 0.05
instead of 0 as mentioned previously in order to ensure
no analyte peaks are incorporated into the noise estimation.
The upper limit is set at 0.2 so that it does not
interact with any potentially doubly charged peaks, or the
"echo" of those doubly charged peaks. Both of these values
can be changed if desired by the user.
}
\examples{
KMDNoise(df, upper = 0.2, lower = 0.05)
KMDNoise(df)

}
