% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{LowMoles}
\alias{LowMoles}
\title{Minimum limit of moles (user limit)}
\usage{
LowMoles(x)
}
\arguments{
\item{x}{component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})}
}
\value{
numeric; value of minimum limit of moles
}
\description{
Returns the minimum limit of moles used by the
  \code{\link{MFAssign}} function, specified by the user.
}
\details{
This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
LowMoles("C")

LowMoles('C')

}
