% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Halogen.R
\name{LowMoles2}
\alias{LowMoles2}
\title{#' Minimum limit of moles (program limit)
#'
#' Returns the minimum limit of moles used by the
#'   \code{\link{MFAssign}} function.
#'
#'   This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
#'  environment.
#'
#' @param x component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})
#'
#' @return numeric; value of minimum limit of moles
#'
#' @examples
#' MinMoles("C")
#'
#' MinMoles('C')
#'
#' @export
#'
MinMoles2 <- function(x) {
  # MinMoles - program defined minimum moles of components,
  # input:  factor {"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"}
  # output: numeric
  y <- 0
  switch(as.character(x),
         "C"  = {y <- 1},
         "H"  = {y <-  2},
         "O"  = {y <- 0},
         "N"  = {y <- 0},
         "S"  = {y <- 0},
         "P"  = {y <- 0},
         "M"  = {y <- 0},
         "E"  = {y <- 0},
         "Z"  = {y <- 1},
         "Cl" = {y <- 0},
         "Fl" = {y <- 0},
         "POE" = {y <- 0},
         "NOE" = {y <- 0},
         "Cl37" = {y <- 0},
         "S34" = {y <- 0},
         "C13" = {y <- 0},
         "Br" = {y <- 0},
         "Br81" = {y <- 0},
         "I"  = {y <- 0},
         { stop("MinMoles called on undefined component ") })
  return(y)
}}
\usage{
LowMoles2(x)
}
\arguments{
\item{x}{component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})}
}
\value{
numeric; value of minimum limit of moles
}
\description{
#' Maximum limit of moles (program limit)
#'
#' Returns the maximum limit of moles used by the
#'   \code{\link{MFAssign}} function.
#'
#'   This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
#'  environment.
#'
#' @param x component; factor (\code{"C", "H", "O", "N", "S", "P", "M", "E", "Z"})
#'
#' @return numeric; value of maximum limit of moles
#'
#' @examples
#' MaxMoles("C")
#'
#' MaxMoles('C')
#'
#' @export
#'
MaxMoles2 <- function(x) {
  # MaxMoles - program maximum limit of moles
  # input:  factor {"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"}
  # output: numeric
  y <- 0
  switch(as.character(x),
         "C"  = {y <- 166},
         "H"  = {y <- 284},
         "O"  = {y <-  72},
         "N"  = {y <-   24},
         "S"  = {y <-   8},
         "P"  = {y <-   5},
         "M"  = {y <-   2},
         "E"  = {y <-   3},
         "Z"  = {y <-   5},
         "Cl" = {y <-   5},
         "Fl" = {y <-   5},
         "POE" = {y <-   1},
         "NOE" = {y <-   1},
         "Cl37" = {y <- 3},
         "S34" = {y <-  3},
         { stop("MaxMoles called on undefined component ") })
  return(y)
}
}
\details{
Minimum limit of moles (user limit)

Returns the minimum limit of moles used by the
  \code{\link{MFAssign}} function, specified by the user.

  This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
LowMoles("C")

LowMoles('C')

}
