% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{merge_and_filter_abundances}
\alias{merge_and_filter_abundances}
\title{Merges and filters abundances for mono and iso pairs}
\usage{
merge_and_filter_abundances(final_pair2, end_data)
}
\arguments{
\item{final_pair2}{data frame:
The data frame containing columns "Mono_mass", "Iso_mass1", and "Iso_mass2".}

\item{end_data}{data frame:
The data frame containing columns "Exp_mass", "Abundance", "RT".}
}
\value{
data frame:
The merged and filtered data frame containing columns "Mono_mass," "Mono_Abund," "Mono_RT," "Iso_mass1," "Iso_Abund1," "Iso_RT1," "Iso_mass2," "Iso_Abund2," and "Iso_RT2."
}
\description{
This function merges mono and iso pairs with abundance data and filters the result.
}
