% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{merge_and_filter_pairs}
\alias{merge_and_filter_pairs}
\title{Merges and filters mono and iso pairs}
\usage{
merge_and_filter_pairs(mono_pair, iso_pair)
}
\arguments{
\item{mono_pair}{data frame:
The data frame containing columns "Mono_mass" and "Iso_mass1."}

\item{iso_pair}{data frame:
The data frame containing columns "Iso_mass1" and "Iso_mass2."}
}
\value{
data frame:
The merged and filtered data frame containing columns "Mono_mass," "Iso_mass1," and "Iso_mass2."
}
\description{
This function merges mono and iso pairs based on the "Iso_mass1" column and filters the result.
}
