% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFAssign.R
\name{MFAssignAll}
\alias{MFAssignAll}
\title{Assigns all possible MF to each row of input data frame with CHOFIT algorithm}
\usage{
MFAssignAll(peaks, isopeaks = "None", ionMode, lowMW = 100, highMW = 1000,
  POEx = 0, Nx = 0, Sx = 0, Px = 0, S34x = 0, N15x = 0, Dx = 0,
  Ex = 0, Clx = 0, Cl37x = 0, Mx = 0, NH4x = 0, Zx = 1, Ox = 30,
  ppm_err = 3, SN = 0, O_Cmin = 0, O_Cmax = 2.5, H_Cmin = 0.3,
  H_Cmax = 3, DBEOmin = -13, DBEOmax = 13, Omin = 0, HetCut = "off",
  NMScut = "on")
}
\arguments{
\item{peaks}{data frame, Monoisotopic Masses}

\item{isopeaks}{data frame, Isotopic Masses}

\item{ionMode}{string: ("neg", "pos")}

\item{lowMW}{numeric:
Sets the lower limit of molecular mass to be assigned. Default is 100.}

\item{highMW}{numeric:
Sets the upper limit of molecular mass to be assigned. Default is 1000.}

\item{POEx}{numeric:
If set to 1 and ionMode is positive, positive mode odd electron ions can be assigned.
Default is 0}

\item{Nx}{numeric:
Sets the maximum allowable number of Nitrogen 14 to be used in assignment. Default is 0.}

\item{Sx}{numeric:
Sets the maximum allowable number of Sulfur 32 to be used in assignment. Default is 0.}

\item{Px}{numeric:
Sets the maximum allowable number of Phosphorus 31 to be used in assignment. Default is 0.}

\item{S34x}{numeric:
Sets the amount of Sulfur 34 to be used in assignment. Default is 0.}

\item{N15x}{numeric:
Sets the amount of Nitrogen 15 to be used in assignment. Default is 0.}

\item{Dx}{numeric:
Sets the amount of Deuterium to be used in assignment. Default is 0.}

\item{Ex}{numeric:
Sets the amount of Carbon 13 to be used in assignment. Default is 0.}

\item{Clx}{numeric:
Sets the amount of Chlorine to be used in assignment. Default is 0.}

\item{Cl37x}{numeric:
Sets the amount of Chlorine 37 to be used in assignment. Default is 0.}

\item{Mx}{numeric:
Sets the amount of Sodium adduct to be used in assignment. Default is 0.}

\item{NH4x}{numeric:
Sets the amount of Ammonium adduct to be used in assignment. Default is 0.}

\item{Zx}{numeric:
Sets the amount of charge to be used in assignment. Default is 1.}

\item{Ox}{numeric:
Ox sets the maximum number of oxygen looked for in the CHOFIT core, it limits the number of loops performed}

\item{ppm_err}{numeric:
ppm_err parameter sets the error tolerance (ppm) for formula assignment. Default is 3.}

\item{SN}{numeric:
SN parameter set the signal to noise cut for formula assignment. Default is 0.}

\item{O_Cmin}{numeric:
O_Cmin parameter sets the minimum allowed oxygen to carbon ratio. Default is 0.}

\item{O_Cmax}{numeric:
The O_Cmax parameter sets the upper limit for oxygen to carbon ratio. Default is 2.5.}

\item{H_Cmin}{numeric:
H_Cmin parameter sets lower limit for hydrogen to carbon ratio. Default is 0.1.}

\item{H_Cmax}{numeric:
H_Cmax parameter sets upper limit for hydrogen to carbon ratio for assigned formulas.
Default is 3.}

\item{DBEOmin}{numeric:
DBEOmin parameter sets lower limit for DBE minus oxygen QA parameter. Default is -13.}

\item{DBEOmax}{numeric:
DBEOmax parameter sets upper limit for DBE minus oxygen QA parameter. Default is 13.}

\item{Omin}{numeric:
Omin parameter sets lower limit for oxygen number in assigned formula. Default is 0.}

\item{HetCut}{character:
HetCut turns on or off the high heteroatom QA parameter. Default is "off"}

\item{NMScut}{character:
NMScut turns on or off the nominal mass series QA parameter. Default is “on”.}
}
\value{
list(Unambig = Unambig, Ambig = Ambigout, None = unassigned, MSAssign = MZ,
         Error = Error, MSgroups = MZgroups, VK = VK)

  Unambig - data frame containing unambiguous assignments
  Ambig - data frame containing ambiguous assignments
  None - data frame containing unassigned masses
  MSAssign - ggplot of mass spectrum highlighting assigned/unassigned
  Error - ggplot of the Error vs. m/z
  MSgroups - ggplot of mass spectrum colored by molecular group
  VK - ggplot of van Krevelen plot, colored by molecular group
}
\description{
MFAssign assigns all possible molecular formulae to each
mass in the input file, subject to user constraints on the moles of
C, H, O, N, S, P, sodium (M), C13 (E), S34, N15, Deuterium (D), Cl, Cl37, NH4, and Z.
}
\details{
There are user inputs for heteroatoms, adducts, and charge.
The terms for each are Nx, Sx, Px, Ex, S34x, N15x, Dx,
Clx, Cl37x, Mx, NH4x, and Zx. Basic QA steps are included within the
function. More detail about these QA steps can be seen in the
vignette and user manual attached to this package.
Additionally, an option to remove ambiguous assignments based on choosing
the formula with the fewest number of heteroatoms (Ohno and Ohno, 2013), and
the CH4 vs O replacement nominal mass series (Koch et al., 2007) is included.
Most of these QA parameters can be adjusted from outside the function.

Positive mode odd-electron formula assignment is possible using the POEx
parameter. If set to 1 positive mode odd-electron ionized masses can be
assigned. This is only really needed for APCI and APPI ionization modes.
This option is currently unavailable, but will be addressed in a future release.

The Ox term sets the number of loops performed by the CHOFIT core by defining the
maximum oxygen allowed to be looked for. Three oxygen is equal to one loop due to
the low mass moiety used in CHOFIT (Perdue and Green 2015). The default setting is
30 oxygen, so it is limited to 10 loops. It can be increased if necessary.

The input dataframe should have two columns, the first column should contain the
relative abundance or intensity for the ion mass and the second column should contain
the ion mass, either positive or negative mode. The function can also handle an additional
input dataframe containing the masses identified as istopic by the IsoFiltR function. The
outputs of that function can be directly put into this function as peaks and isopeaks,
doing so will allow likely 13C isotope masses to be match to assigned monoisotopic masses.
The inclusion of an isotopic mass list is not required, however.
}
\examples{
MFAssign(peaks = Mono_df, isopeaks = Iso_df, "neg", lowMW = 200, highMW = 700)
MFAssign(peaks = Mono_df, isopeaks = Iso_df, "neg", lowMW = 100, highMW = 1000, Nx = 3, Sx = 1)
}
