% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recal Functions.R
\name{MFRecalCheck}
\alias{MFRecalCheck}
\title{Generates a plot to check chosen recalibrant series, recalibrates two mass lists,
and produces a list of the chosen recalibrants.}
\usage{
MFRecalCheck(df, peaks, isopeaks = "None", mode, SN = 0, series1 = NA,
  series2 = NA, series3 = NA, series4 = NA, series5 = NA,
  series6 = NA, series7 = NA, series8 = NA, series9 = NA,
  series10 = NA, min = 100, max = 1000, bin = 20, obs = 2)
}
\arguments{
\item{df}{data frame output from MFAssign() or MFAssignCHO()}

\item{peaks}{data frame with two columns for recalibration, generally contains "Mono" output from IsoFiltR()}

\item{isopeaks}{data frame with two columns, generally containing the "Iso" output from IsoFiltR()}

\item{mode}{character "neg" or "pos" depending on the ionization mode of the data}

\item{SN}{numeric value of the signal-to-noise cut point for making the plot, default is 0}

\item{min}{numeric - minimum mass range of the data being analyzed.}

\item{max}{numeric - maximum mass range of the data being analyzed.}

\item{bin}{numeric - sets the mass window bins to choose recalibrants from, default is 20.}

\item{obs}{numeric - sets the number of peaks to choose as recalibrants from each bin,
default is 2}

\item{series(1-10)}{character the recalibrant series, "O7_Na_4" for example, default is NA}
}
\value{
list(MZ, Mono, Iso, RecalOut) contains the mass spectrum (MZ), recalibrated "Mono" data frame,
 recalibrated "Iso" data frame, and the recalibrants list (RecalOut)
}
\description{
MFRecalCheck() takes the data frame output from MFAssign, the outputs of IsoFiltR(), and
the chosen recalibrant series to generate a plot for the qualitative assessment of the recalibrants,
recalibrate the "Mono" and "Iso" outputs from IsoFiltR, and prepare a data frame containing the chosen
recalibrants.
}
\details{
This function can handle up to 10 homologous series, though it will work with fewer. It is important
for recalibrant masses to cover the entire mass range of interest, and they should be among the
most abundant peaks in their region of the spectrum, this function helps with choosing and
visualizing the recalibrants.

This function also recalibrates up to two mass lists using the chosen recalibrants.
It is best to use the "Mono" and "Iso" outputs of the IsoFiltR() function.

These parameters will be used to identify more potential recalibrants, and the tallest peaks within
a user defined mass range will be used as recalibrants.
}
\examples{
MFRecalCheck(df = Data, peaks = Mono, isopeaks = Iso, mode = "neg", SN = 500, series1 = "O4_H_2", series2 = "O4_H_8", series3 = "O6_H_8")

MFRecalCheck(df = Data, peaks = Mono, isopeaks = Iso, mode = "pos", SN = 300, series1 = "O4_Na_2", series2 = "O4_H_8", series3 = "O6_Na_8")
}
