% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities sks.R
\name{MinMoles}
\alias{MinMoles}
\title{Minimum limit of moles (program limit)}
\usage{
MinMoles(x)
}
\arguments{
\item{x}{component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})}
}
\value{
numeric; value of minimum limit of moles
}
\description{
Returns the minimum limit of moles used by the
  \code{\link{MFAssign}} function.
}
\details{
This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
MinMoles("C")

MinMoles('C')

}
