% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecalFunctions.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot Spectrum Function}
\usage{
plot_spectrum(plotpeak, RecalPlot)
}
\arguments{
\item{plotpeak}{A data frame containing information for the first set of segments.}

\item{RecalPlot}{A data frame containing information for the second set of segments.}
}
\value{
A ggplot object representing the mass spectrum plot.
}
\description{
This function generates a mass spectrum plot using ggplot2.
}
