% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{process_carbon_isotoping}
\alias{process_carbon_isotoping}
\title{Process carbon isotoping data}
\usage{
process_carbon_isotoping(raw_data, carb_error, carb_ratio, end_data)
}
\arguments{
\item{raw_data}{data frame:
Raw data containing information for carbon isotoping.}

\item{carb_error}{numeric:
Error for carbon isotoping.}

\item{carb_ratio}{numeric:
Carbon ratio for filtering.}

\item{end_data}{data frame:
End data containing information for carbon isotoping.}
}
\value{
list:
A list containing three data frames: "MonoC" for monoisotopic data, "IsoC1" for first isotope data, and "IsoC2" for second isotope data.
}
\description{
This function processes carbon isotoping data using specified parameters and filtering criteria.
}
