% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{process_sulfur}
\alias{process_sulfur}
\title{Process sulfur data}
\usage{
process_sulfur(raw_data, sulferr, data_end, sulfrat)
}
\arguments{
\item{raw_data}{data frame:
Raw data containing information for sulfur.}

\item{sulferr}{numeric:
Error for sulfur.}

\item{data_end}{data frame:
End data containing information for sulfur.}

\item{sulfrat}{numeric:
Sulfur ratio for filtering.}
}
\value{
list:
A list containing two data frames: "IsooutS" for isotope data and "MonooutS" for monoisotopic data.
}
\description{
This function processes sulfur data using specified parameters and filtering criteria.
}
