% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecalFunctions.R
\name{processKnown}
\alias{processKnown}
\title{Process Known Compounds}
\usage{
processKnown(
  rest,
  known,
  kmd_col,
  z_col,
  num_col,
  type,
  step_limit,
  remove_indices
)
}
\arguments{
\item{rest}{The dataset containing the rest of the compounds.}

\item{known}{The dataset containing known compounds.}

\item{kmd_col}{The column name for KMD in both datasets.}

\item{z_col}{The column name for Z in both datasets.}

\item{num_col}{The column name to store the calculated number of compounds.}

\item{type}{The column name to store the type of compound.}

\item{step_limit}{The limit for considering compounds.}

\item{remove_indices}{Indices of rows to be removed from the result.}
}
\value{
A processed dataset with updated abundance and type columns.
}
\description{
This function processes known compounds by merging them with a rest dataset and
updating abundance and type columns based on specified columns and limits.
}
