% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecalFunctions.R
\name{processMassList}
\alias{processMassList}
\title{Process Mass List}
\usage{
processMassList(data)
}
\arguments{
\item{data}{A data frame containing at least the "Exp_mass" column.}
}
\value{
The input data frame with additional calculated columns.
}
\description{
This function processes a data frame containing mass-related information.
}
\details{
The function calculates the following columns:
\itemize{
\item NM: Rounded Exp_mass
\item KM_O: Exp_mass * (16 / 15.9949146223)
\item KMD_O: Rounded NM - KM_O
\item z_O: Rounded Exp_mass \%\% 16 - 16
\item KM_H2: Exp_mass * (2 / 2.01565)
\item KMD_H2: Rounded NM - KM_H2
\item z_H2: Rounded Exp_mass \%\% 2 - 2
}
}
\examples{
data <- data.frame(Exp_mass = c(100, 150, 200))
processed_data <- processMassList(data)
}
