% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecalFunctions.R
\name{RecalList}
\alias{RecalList}
\title{Identifies canditate series for recalibration}
\usage{
RecalList(df)
}
\arguments{
\item{df}{data frame output from MFAssign() or MFAssignCHO()}
}
\value{
data frame
}
\description{
RecalList() takes the output from MFAssign() and/or MFAssignCHO()
and identifies the homologous series that could be used as recalibrants.
}
\details{
It returns a dataframe that contains the CH2 homologous series that
contain more than 3 members.

The columns of the returned dataframe are as follows:

Series - reports the homologous series according to class, adduct, and DBE,
the format is "class_adduct_DBE", for example a homologous series with
class = "O6, adduct of Na+, and DBE = 4 would be "O6_Na_4"

Number Observed - reports the number of members of each homologous series.

Series Index - represents the order of the series when ordered by length of
homologous series.

Mass Range - reports the minimum and maximum mass for the compounds within a
homologous series.

Tall Peak - reports the mass of the most abundant peak in each series.
Abundance Score - reports the percentage difference between the mean abundance
of a homologous series and the median abundance within the mass range the
"Tall Peak" falls in (for example m/z 200-300). A higher score is generally better.

Peak Score - This column compares the intensity of the tallest peak in a
given series to the second tallest peak in the series This comparison is
calculated by log10(Max Peak Intensity/Second Peak Intensity) The closer
to 0 this value is the better, in general.

Peak Distance - This column shows the number of CH2 units between the tallest
and second tallest peak in each series. In general, it is better for the
value to be as close to 1 as possible.

Series Score - This column compares the number of actual observations in each
series to the theoretical maximum number based on the CH2 homologous series.
The closer to one this value is, the better.

While the function does make some minor decisions as to which series to choose
from (more than 3 members of homologous series) the rest of the decision
making is left to the user.
}
\examples{
RecalList(df = Data)
}
