% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNcutCheck.R
\name{SNcutCheck}
\alias{SNcutCheck}
\title{Estimates the Signal to Noise cut}
\usage{
SNcutCheck(df, SN = 0, bin = 0.01)
}
\arguments{
\item{df}{- dataframe of intensity and ion mass, column 1 should be intensity, column 2 should be mass}

\item{SN}{- numeric value for situations where a predefined S/N value is desired, default is 0}

\item{bin}{- numeric value determining the binwidth of the histogram, default is 0.01}
}
\value{
S/N and log intensity histogram
  S/N printed in Console
  Histogram shows where the cut is being applied
}
\description{
For raw mass spectral data using the method
developed by Zhurov et al. Anal. Chem. (2014).
It uses the natural log of the intensity to create a histogram which then
is used to determine the S/N cut for the data being used.
The data being put into this function should be the raw mass list output from
a mass spectrum, if the noise peaks have been removed already, it will not work.
}
\details{
The value reported in the console will be the estimated S/N cut and the plot
will demonstrate where the cut is being applied to the data. The value in the console
can then be multiplied by whatever value is desired in order to reach the value
to be used to cut the data.
}
\examples{
SNcutCheck(df, SN = 0, bin = 0.01)
SNcutCheck(df)

}
