% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoFiltR.R
\name{sulfur_check}
\alias{sulfur_check}
\title{Check sulfur data}
\usage{
sulfur_check(MonoOutC_final, MonoOutS_final, IsoOutC1_final, IsoOutC2_final)
}
\arguments{
\item{MonoOutC_final}{data frame:
Data frame containing final monoisotopic carbon data.}

\item{MonoOutS_final}{data frame:
Data frame containing final monoisotopic sulfur data.}

\item{IsoOutC1_final}{data frame:
Data frame containing final isotope 1 carbon data.}

\item{IsoOutC2_final}{data frame:
Data frame containing final isotope 2 carbon data.}
}
\value{
list:
A list containing three processed data frames: monoisotopic sulfur, isotope 1 carbon, and isotope 2 carbon.
}
\description{
This function performs sulfur-related checks and processes the data accordingly.
}
