% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Halogen.R
\name{Valence2}
\alias{Valence2}
\title{Valence of component}
\usage{
Valence2(x)
}
\arguments{
\item{x}{component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})}
}
\value{
numeric; value of valence
}
\description{
Returns the valence of the specified component; used by the
  \code{\link{MFAssign}} function.
 This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
Valence("C")

Valence('C')

}
