% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities sks.R
\name{ValidFormula}
\alias{ValidFormula}
\title{Determine if molecular formula is valid}
\usage{
ValidFormula(moles)
}
\arguments{
\item{moles}{vector of integers (length 9) - indicating number of each
elements}
}
\value{
boolean
}
\description{
Evaluates if a molecular formula is valid, obeying senior rules and other
constraints
}
\details{
This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
 environment.
}
\examples{
ValidFormula(moles)

ValidFormula(c(2,3,1,0,0,0,0,0,1))

}
