% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities_Halogen.R
\name{LowMoles2}
\alias{LowMoles2}
\title{#' Minimum limit of moles (program limit)
#'
#' Returns the minimum limit of moles used by the
#'   \code{\link{MFAssign}} function.
#'
#'   This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
#'  environment.
#'
#' @param x component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})
#'
#' @return numeric; value of minimum limit of moles
#'
#'
#' @export
#'}
\usage{
LowMoles2(x)
}
\arguments{
\item{x}{component; factor (\code{"C", "H", "O", "N", "S", "P", "E", "S34", "N15", "D", "Cl", "Cl37", "M", "NH4", "Z"})}
}
\value{
numeric; value of minimum limit of moles
}
\description{
#' Maximum limit of moles (program limit)
#'
#' Returns the maximum limit of moles used by the
#'   \code{\link{MFAssign}} function.
#'
#'   This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
#'  environment.
#'
#' @param x component; factor (\code{"C", "H", "O", "N", "S", "P", "M", "E", "Z"})
#'
#' @return numeric; value of maximum limit of moles
#'
#'
#' @export
#'
}
\details{
Minimum limit of moles (user limit)

Returns the minimum limit of moles used by the
\code{\link{MFAssign}} function, specified by the user.

This is an internal fuction that will not work outside the \code{\link{MFAssign}} function
environment.
}
