% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMDNoise.R
\name{SNplot}
\alias{SNplot}
\title{Plot of Mass Spectrum with highlighted S/N cut}
\usage{
SNplot(df, cut, mass, window.x = 0.5, window.y = 10)
}
\arguments{
\item{df}{\itemize{
\item dataframe of intensity and ion mass, column 1 should be intensity, column 2 should be mass
}}

\item{cut}{\itemize{
\item numeric value of the intensity cut value being investigated
}}

\item{mass}{\itemize{
\item numeric value setting a centerpoint to look at the mass spectrum
}}

\item{window.x}{\itemize{
\item numeric value setting the +/- range around the mass centerpoint, default is 0.5
}}

\item{window.y}{\itemize{
\item numeric value setting the y axis value for the plot, determined by multiplying the cut by this value
}}
}
\value{
S/N cut colored mass spectrum

Zoomed Mass spectrum which shows where the cut is being applied
}
\description{
SNplot which plots the mass spectrum with the S/N cut denoted by different colors
for the mass spectrum peaks. This is useful for a qualitative look at the effectiveness
of the S/N cut being used.
}
